RECEIVE_CHART_TYPE middleware
